Isolate external review signal collection
