+++ .codex-supervisor/issue-journal.md
@@ -1,37 +1,39 @@
-# Issue #489: Test cleanup: split supervisor PR lifecycle coverage by readiness and review-blocker boundaries
+# Issue #491: External-review cleanup: tighten the signal-heuristics and normalization boundary
 
 ## Supervisor Snapshot
-- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/489
-- Branch: codex/issue-489
-- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-489
-- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-489/.codex-supervisor/issue-journal.md
+- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/491
+- Branch: codex/issue-491
+- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-491
+- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-491/.codex-supervisor/issue-journal.md
 - Current phase: reproducing
 - Attempt count: 1 (implementation=1, repair=0)
-- Last head SHA: 9704168f31e53da4a7139911a933b74613c5ca4c
+- Last head SHA: 18af0964675dd972a9db716197af125cf9b75cba
 - Blocked reason: none
 - Last failure signature: none
 - Repeated failure signature count: 0
-- Updated at: 2026-03-17T15:54:52.619Z
+- Updated at: 2026-03-17T16:15:30.271Z
 
 ## Latest Codex Summary
-- Split `src/supervisor/supervisor-pr-lifecycle.test.ts` into readiness and review-blocker suites, preserved the legacy file as a thin import facade, and re-verified the focused supervisor PR lifecycle tests plus `npm run build`.
+- None yet.
 
 ## Active Failure Context
 - None recorded.
 
 ## Codex Working Notes
 ### Current Handoff
-- Hypothesis: #489 can stay behavior-neutral by splitting `supervisor-pr-lifecycle.test.ts` into one readiness/post-turn suite and one review-blocker/thread-reprocessing suite, with the legacy path preserved as a facade import for existing test entry points.
-- What changed: extracted readiness coverage into `src/supervisor/supervisor-pr-readiness.test.ts`, moved review-blocker and configured-bot thread reprocessing coverage into `src/supervisor/supervisor-pr-review-blockers.test.ts`, and reduced `src/supervisor/supervisor-pr-lifecycle.test.ts` to a thin facade import file.
+- Hypothesis: #491 can stay behavior-neutral by moving provider-specific external-review signal collection out of `external-review-normalization.ts`, leaving normalization focused on shaping preselected signal envelopes into findings.
+- What changed: added focused `external-review-signal-collection.test.ts` coverage for provider-activity filtering, extracted `collectExternalReviewSignals` and the thread/review/comment selection helpers into `src/external-review/external-review-signal-collection.ts`, moved shared signal types into `src/external-review/external-review-signals.ts`, and updated normalization/classifier/miss-persistence tests to use the explicit `signal -> normalized finding` boundary.
 - Current blocker: none
-- Next exact step: Commit the suite split on `codex/issue-489`, then push/open the draft PR if one still does not exist.
-- Verification gap: none for the split PR lifecycle suites or `npm run build`; `npm ci` was required first because `tsc` was initially unavailable locally.
-- Files touched: `src/supervisor/supervisor-pr-lifecycle.test.ts`, `src/supervisor/supervisor-pr-readiness.test.ts`, `src/supervisor/supervisor-pr-review-blockers.test.ts`, `.codex-supervisor/issue-journal.md`
-- Rollback concern: reverting this cleanup would put readiness transitions, post-turn PR refresh, configured-bot thread reprocessing, and manual/local review blockers back into a single 915-line suite, making failures harder to localize.
+- Next exact step: Commit the external-review boundary cleanup on `codex/issue-491`, then check whether a draft PR already exists for the branch before opening or updating one.
+- Verification gap: none for the focused external-review suites or `npm run build`; `npm ci` was required first because `tsc` was initially unavailable locally in this worktree.
+- Files touched: `src/external-review/external-review-signal-collection.ts`, `src/external-review/external-review-signals.ts`, `src/external-review/external-review-normalization.ts`, `src/external-review/external-review-misses.ts`, `src/external-review/external-review-miss-persistence.ts`, `src/external-review/external-review-durable-guardrail-candidates.ts`, `src/external-review/external-review-miss-artifact-types.ts`, `src/external-review/external-review-signal-collection.test.ts`, `src/external-review/external-review-normalization.test.ts`, `src/external-review/external-review-classifier.test.ts`, `src/external-review/external-review-miss-persistence.test.ts`, `.codex-supervisor/issue-journal.md`
+- Rollback concern: reverting this cleanup would put configured-bot selection and actionable filtering back into normalization, which makes future provider-signal changes span both collection and finding-shaping code paths again.
 - Last focused command: `npm run build`
 ### Scratchpad
-- 2026-03-18: Focused reproducer for #489 was `npx tsx --test src/supervisor/supervisor-pr-lifecycle.test.ts`, which passed before the split and confirmed the cleanup was behavior-neutral rather than a failing behavior fix.
-- 2026-03-18: Verification for #489 was `npx tsx --test src/supervisor/supervisor-pr-readiness.test.ts src/supervisor/supervisor-pr-review-blockers.test.ts`, `npx tsx --test src/supervisor/supervisor-pr-lifecycle.test.ts`, then `npm ci` and `npm run build`; `npm run build` first failed with `sh: 1: tsc: not found` before dependencies were restored.
+- 2026-03-18: Focused reproducer for #491 was a new `external-review-signal-collection.test.ts` suite that pins provider-activity behavior separately from `normalizeExternalReviewSignal`, including last-configured-bot thread comment selection, actionable state-only top-level reviews, and closed-PR issue-comment filtering.
+- 2026-03-18: #491 cleanup extracts provider signal collection into `external-review-signal-collection.ts` and shared signal types into `external-review-signals.ts`; `external-review-normalization.ts` now only shapes signal envelopes into normalized findings.
+- 2026-03-18: Focused verification for #491 was `npx tsx --test src/external-review/external-review-signal-collection.test.ts src/external-review/external-review-normalization.test.ts src/external-review/external-review-classifier.test.ts src/external-review/external-review-miss-persistence.test.ts`.
+- 2026-03-18: `npm run build` initially failed with `sh: 1: tsc: not found`; `npm ci` restored the local toolchain, then the rerun of the focused suites and `npm run build` passed.
 - 2026-03-18: `npm run build` initially failed with `sh: 1: tsc: not found`; `npm ci` restored the local toolchain and the rerun of `npm run build` passed.
 - 2026-03-17: Review repair for PR #497 updates the guardrail provenance fixtures so committed shared-memory files are added before `headSha` is captured; focused verification was `npx tsx --test src/supervisor/supervisor-diagnostics-guardrail-reporting.test.ts` and `npm run build`.
 - 2026-03-17: Added `draftSkipAt` to configured-bot summaries and hydrated PRs; focused verification was `npx tsx --test src/github/github-review-signals.test.ts src/github/github-pull-request-hydrator.test.ts`, followed by `npm ci` and `npm run build`.
