+++ src/external-review/external-review-normalization.ts
@@ -1,20 +1,6 @@
 import { truncate } from "../core/utils";
 import { type LocalReviewSeverity } from "../local-review/types";
-import { hasActionableReviewText, isActionableTopLevelReview } from "./external-review-signal-heuristics";
-import { type IssueComment, type PullRequestReview, type ReviewThread } from "../core/types";
-
-export type ExternalReviewSignalSourceKind = "review_thread" | "top_level_review" | "issue_comment";
-
-export interface ExternalReviewSignalEnvelope {
-  sourceKind: ExternalReviewSignalSourceKind;
-  sourceId: string;
-  sourceUrl: string | null;
-  reviewerLogin: string;
-  body: string;
-  file: string | null;
-  line: number | null;
-  threadId: string | null;
-}
+import { type ExternalReviewSignalEnvelope, type ExternalReviewSignalSourceKind } from "./external-review-signals";
 
 export interface NormalizedExternalReviewFinding {
   source: "external_bot";
@@ -72,112 +58,6 @@ function inferConfidence(body: string): number {
   return 0.75;
 }
 
-function normalizeBotLogins(reviewBotLogins: string[]): Set<string> {
-  return new Set(reviewBotLogins.map((login) => login.toLowerCase()));
-}
-
-function latestConfiguredBotComment(thread: ReviewThread, allowed: Set<string>) {
-  for (let index = thread.comments.nodes.length - 1; index >= 0; index -= 1) {
-    const comment = thread.comments.nodes[index];
-    const login = comment.author?.login?.toLowerCase();
-    if (login && allowed.has(login)) {
-      return comment;
-    }
-  }
-
-  return null;
-}
-
-export function toExternalReviewThreadSignal(
-  thread: ReviewThread,
-  reviewBotLogins: string[],
-): ExternalReviewSignalEnvelope | null {
-  const comment = latestConfiguredBotComment(thread, normalizeBotLogins(reviewBotLogins));
-  if (!comment) {
-    return null;
-  }
-
-  return {
-    sourceKind: "review_thread",
-    sourceId: thread.id,
-    sourceUrl: comment.url ?? null,
-    reviewerLogin: comment.author?.login ?? "unknown",
-    body: comment.body,
-    file: thread.path ?? null,
-    line: thread.line ?? null,
-    threadId: thread.id,
-  };
-}
-
-export function toExternalTopLevelReviewSignal(
-  review: PullRequestReview,
-  reviewBotLogins: string[],
-): ExternalReviewSignalEnvelope | null {
-  const login = review.author?.login?.toLowerCase();
-  if (!login || !normalizeBotLogins(reviewBotLogins).has(login) || !isActionableTopLevelReview(review)) {
-    return null;
-  }
-
-  const body = review.body?.trim() || review.state?.trim();
-  if (!body) {
-    return null;
-  }
-
-  return {
-    sourceKind: "top_level_review",
-    sourceId: review.id,
-    sourceUrl: review.url ?? null,
-    reviewerLogin: review.author?.login ?? "unknown",
-    body,
-    file: null,
-    line: null,
-    threadId: null,
-  };
-}
-
-export function toExternalIssueCommentSignal(
-  comment: IssueComment,
-  reviewBotLogins: string[],
-): ExternalReviewSignalEnvelope | null {
-  const login = comment.author?.login?.toLowerCase();
-  if (!login || !normalizeBotLogins(reviewBotLogins).has(login) || !hasActionableReviewText(comment.body)) {
-    return null;
-  }
-
-  return {
-    sourceKind: "issue_comment",
-    sourceId: comment.id,
-    sourceUrl: comment.url ?? null,
-    reviewerLogin: comment.author?.login ?? "unknown",
-    body: comment.body,
-    file: null,
-    line: null,
-    threadId: null,
-  };
-}
-
-export function collectExternalReviewSignals(args: {
-  reviewThreads?: ReviewThread[];
-  reviews?: PullRequestReview[];
-  issueComments?: IssueComment[];
-  reviewBotLogins: string[];
-}): ExternalReviewSignalEnvelope[] {
-  return [
-    ...(args.reviewThreads ?? []).flatMap((thread) => {
-      const signal = toExternalReviewThreadSignal(thread, args.reviewBotLogins);
-      return signal ? [signal] : [];
-    }),
-    ...(args.reviews ?? []).flatMap((review) => {
-      const signal = toExternalTopLevelReviewSignal(review, args.reviewBotLogins);
-      return signal ? [signal] : [];
-    }),
-    ...(args.issueComments ?? []).flatMap((comment) => {
-      const signal = toExternalIssueCommentSignal(comment, args.reviewBotLogins);
-      return signal ? [signal] : [];
-    }),
-  ];
-}
-
 export function normalizeExternalReviewSignal(
   signal: ExternalReviewSignalEnvelope,
 ): NormalizedExternalReviewFinding | null {
@@ -201,14 +81,6 @@ export function normalizeExternalReviewSignal(
     confidence: inferConfidence(signal.body),
     url: signal.sourceUrl,
   };
-}
-
-export function normalizeExternalReviewFinding(
-  thread: ReviewThread,
-  reviewBotLogins: string[],
-): NormalizedExternalReviewFinding | null {
-  const signal = toExternalReviewThreadSignal(thread, reviewBotLogins);
-  return signal ? normalizeExternalReviewSignal(signal) : null;
 }
 
 export function createExternalReviewMissPatternFingerprint(
